split out fstabentry code into a seperate module
in hdimg.py parted is importet and fstabentry is defined. But the code
that uses the fstabentry class, doesn't need parted. fstabentry is used in
fstab.mako, so python-parted was necessary on the host sytem for no real
technical reason. This patch moves the fstabentry code into a seperate module
and copies it to the target image, etc.

Signed-off-by: Manuel Traut <manut@linutronix.de>