add target pkg list support
The current pkg list is the list that belongs to the buildenv. If a special
copy mode or finetuning is used, the list of the target may be shorter.

This patch adds a new feature to add a target package list to the target
rootfile-system. This is useful if dpkg is not available on the target and
exporting versions, e.g. in a webinterface is needed.

To add the list to the target the pkgversionlist tag needs to be set in the
XML file. The file on the target is called /etc/elbe_pkglist

The target package list is also added to elbe-report.txt

Signed-off-by: Manuel Traut <manut@linutronix.de>