enable local debian repositories for use with elbe
if you want to use a debian repository on your local machine (the box you
are typing the elbe commands). You can now use sth. like
deb http://LOCALMACHINE/my-repo wheezy main
in the url-list.

LOCALMACHINE will be replaced by localhost for elbe create and
buildchroot.py. For the debian-installer and inside the buildenv 10.0.2.2
will be used. This is the IP of your local machine can be reached inside
qemu.

Signed-off-by: Manuel Traut <manut@linutronix.de>