fix diet mode for wheezy
wheezy has multilib support, so the name of the architecture was added to the
list and conffiles in /var/lib/dpkg/info

the extension isn't used for most of the packages including binaries. So this
patch adds a 2nd line to the shell script that also looks for a file including
the architecture name.

Signed-off-by: Manuel Traut <manut@linutronix.de>