adjustpkgs: mark "unknown" packages as auto instead of deleting them
As of now it marks certain packages as "delete" like libgnutls26. This
breaks the wget package (because it depends on the former) and removes
it then.
With this change, we don't remove any packages anymore as this mark_auto
seems to do nothing. This also looks like the incremental updates
(elbe buildchroot --skip-debootrap) do no longer remove packages which
were removed from the xml file.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>