create-target-rfs: add a root prefix
usually we get chrooted into our new rfs and create there /target within
it where we copy the new rfs. This patch makes it possible to work
without the chroot and specify the root directory. By default nothing
changes and it works like before i.e.
 -r / --target /
With this change one can specify the target directory to be outside of
the freshly created filesystem
 -r build/chroot/ -target buld/target/

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>