ElbeProject: Add set_xml to change XML file
This adds a new method set_xml to the ElbeProject class to change the
XML file in an existing ElbeProject instance. The method checks if the
architecture of the new XML file matches the old one and throws an error
otherwise. After changing the XML file, we also need to reinitialise the
buildenv, targetfs and _rpcaptcache members, if we had them before. For
_rpcaptcache it is enough to set it to None, as the get_rpcaptcache
method will recreate it automatically on first usage.

Signed-off-by: Stefan Gast <stefan.gast@linutronix.de>