ElbeProject: Only clean builddir/target on build
In the constructor of ElbeProject, we need to pass clean=False to the
TargetFs constructor, otherwise builddir/target will be deleted each
time we instanciate ElbeProject. This caused the hdimg command to fail,
because it expects the target directory to be populated by a previous
buildchroot run. With this commit, the old (and correct) behavior is
restored.

In the build method, we always have to create a new TargetFs instance
with clean=True to get a fresh target directory.

In set_xml, we do not clean the target directory either, as some changes
to the XML files do not require a rebuild (e.g. changes to the
partitioning scheme), so the directory might be reused in some cases.

Signed-off-by: Stefan Gast <stefan.gast@linutronix.de>