db: Split build_project
Building a project is not a part of the database operations themselves,
so do not start the build process in the database core to keep things
separated cleanly. Instead, provide two database methods
set_build_in_progress and set_build_done to manipulate the project
status in the database. Also adapt elbe db build to work with the new
methods.

Signed-off-by: Stefan Gast <stefan.gast@linutronix.de>