db: Rework error handling, close session when idle
Instead of printing error messages directly to stdout, throw exceptions,
so the users of the db module can decide how to tell the user what went
wrong.

Furthermore, introduce a session_scope for db operations. SQLAlchemy
implicitly organizes db operations in transactions, which are started
upon the first database access and finished when commiting, rolling
back, or closing the session. The session_scope commits any pending
database operations on regular exit. If an exception was raised from
within the session_scope, then a rollback is triggered. Finally, the
session is closed in both cases. This ensures that all database
transactions are terminated correctly. Note that with SQLite, the entire
database is locked during a transaction, see:
http://docs.sqlalchemy.org/en/rel_0_7/dialects/sqlite.html

Signed-off-by: Stefan Gast <stefan.gast@linutronix.de>