Makefile: interpreter/nic/image switch for armhf
armhf needs different qemu parameters for 'debian-installer' run and
make 'run/run-con'.

This is a ugly hack, that implements this in the Makefile.mako. It would
be a lot nicer to introduce different parameters for the different Makefile
targets in xmldefaults. However we're changing everything to a debootstrap
and chroot based buildenvironment for elbe-1.0. So this hack will hopefully
obsolete and removed in the near future.

Signed-off-by: Manuel Traut <manut@linutronix.de>