db: Implement user deletion
Add a function del_user to ElbeDB. The function deletes the user from
the database and sets the owner of all projects of the deleted user to
None. The list of these projects is then returned and can be used for
further cleanup.

The corresponding elbe db command has also been implemented. It
optionally removes all projects that were owned by the user deleted.

Signed-off-by: Stefan Gast <stefan.gast@linutronix.de>