Extend AsyncBuilder to support APT cache commit
An APT cache commit can take some time (depending on the number of
packages to be changed), so add functionality to AsyncBuilder to do this
in the background. Also add a buildtype parameter to
ProjectManager.build_current_project, specifing whether we want to do a
full build or just an APT cache commit.

Signed-off-by: Stefan Gast <stefan.gast@linutronix.de>