db: Increase connection timeout
This is necessary to avoid the occassional "Database is locked" messages
after doing heavy I/O. In particular we had the problem that
set_build_done takes quite long after a build. If we get another query
during the transaction, this can fail as the default timeout is too short.

Signed-off-by: Stefan Gast <stefan.gast@linutronix.de>