use qemu-elbe-user-static if available
qemu-elbe-user-static is packaged by the elbe team and includes a modified
version of qemu-user-static that is installed to /usr/share/elbe/qemu-elbe

it is used to use a patched version of qemu that includes (dirty) workarounds
for elbe that won't be included in upstream or debian.

this code can be removed as soon as the upstream qemu is able to handle
threads and signaling correctly.

however it is a method to use a defined version of qemu-user with elbe.

if the package isn't available or doesn't include a implementation for a
specific architecture, the upstream version is used as a fallback.

Signed-off-by: Manuel Traut <manut@linutronix.de>