Use polymorphism for AsyncWorker jobs
For generating update packages, the AsyncWorker will also need the path
to the XML file of the base version, but there was no way to pass
additional information in AsyncWorkerJobs. So we just eliminate the
buildtype parameter and use polymorphism. The job for generating updates
can then take the path to the base XML file in its constructor, store it
in a member variable and use it in its execute method.

The build_current_project method of the ProjectManager used to take a
buildtype parameter to decide whether to do a build or just an APT cache
commit. The parameter has been removed in favor of a distinct method
apt_commit, which starts the commit job instead (similar to what we
already do with apt_update), so callers have to be updated accordingly.

Signed-off-by: Stefan Gast <stefan.gast@linutronix.de>