docs: Fix target/xmlfile confusion in gen_update
The given XML file is from the base version of the project. The
currently installed projects are compared against the fullpkg list in
the XML file. The update package will upgrade a project created with the
given XML file to the packages currently installed in the given target
directory.

Signed-off-by: Stefan Gast <stefan.gast@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>

Conflicts:
	docs/elbe-gen_update.txt