updatepkg: Always clean update dir. before start
If we don't, an update package can contain package files from an update
generated before, for example if we added the package foo in the previous
update package and now we build an update with the new version being the
base version, then foo is included in the new update package again, even
if it is not needed for the new base version. Also, we should not include
any cruft from failed update package builds. So just delete the
directory to ensure the update package only contains what we want.

Signed-off-by: Stefan Gast <stefan.gast@linutronix.de>