AsyncWorker: Create version DB entry earlier
ElbeDB.save_version contains the check for duplicates, so we should
add the database entry before we create the package archive. Note that
the latter step might fail, so in that case delete the version from the
database again. In fact, we can even create the database entry when
enqueueing the project. This has the further advantage that the user
will see an error message immediately if there already is a version with
the current number.

Signed-off-by: Stefan Gast <stefan.gast@linutronix.de>