updated: use cmd and conf diretories
if one of the above mentioned directory exist, cp the files to the rfs (conf)
  or try to execute all files in the directory (cmd).

For copying the conf files, the according filesystems are remounted
read-writeable (if mounted read-only).

The scripts in cmd need to bother about remounting by their own, because i
can't know what they like to touch.

Signed-off-by: Manuel Traut <manut@linutronix.de>