gpg: define GNUPGHOME variable
If e.g. elbe updated is started by sysvinit GNUPGHOME is set to '$HOME/.gnupg'.
But $HOME is not defined. So this results in using /.gnupg/gnupg.conf.

If it is started by 'root' the $HOME varialbe is set to '/root' and
/root/.gnupg/gnupg.conf is used.

To always look for the config on a defined position the GNUPGHOME environment
variable is now set by elbe to '/var/cache/elbe'.

Signed-off-by: Manuel Traut <manut@linutronix.de>