elbeproject: update mark_delete to work with new cleanup logic
The old cleanup logic needed to be run before package modification.
It marked all packages not in pkglist or debootstrap list as deleted.
This resulted in a new dependency calculation. However the behaviour
was different each run, because the mark_deleted packages influence
the decisions of the apt resolver.

The new cleanup logic needs to run, after the package modifications.
It looks for unused packages and marks them as deleteable.

This patch modifies the mark_delete function the projectmanager to work
with the new logic.

Signed-off-by: Manuel Traut <manut@linutronix.de>