finetuning: fix updated tag to work with foreign architectures
we need to execute the rpcaptcache in the buildenv chroot to avoid
getting strange errors, because the python-apt bindings try to
execve some apt binaries. These binaries (e.g. /usr/lib/apt/methods/http)
are dynamicly linked and need to be executed by qemu-* inside the chroot
if a foreign architecture is build.

Signed-off-by: Manuel Traut <manut@linutronix.de>