hdimg: call update-initramfs before installing grub
update-initramfs is called when the kernel is installed.
It scans /etc/fstab for the filesystem and then places the
right fsck.* executable into the initramfs.

However, /etc/fstab is not yet written, when the kernel package is
installed. We should actually write /etc/fstab before installing
packages, but we shy away from the potential harm this would do,
and just regenerate the initramfs when grub is installed.

Signed-off-by: Torben Hohn <torbenh@linutronix.de>