elbe chg_archive: use .tbz or a directory
Often the archive is maintained in a vcs to track the changes.
Then it was complicated and error prone to create a tarball from
this archive directory and set it to the XML file.

Now this directory can be passed instead of the intermediate
tbz directly.

If the parameter --keep-attributes is passed. The local file
permissions are kept. If not, everything belongs to root:root.

Signed-off-by: Manuel Traut <manut@linutronix.de>