Update dependencies of updated for usb monitor.
As the usb monitor uses udev to monitor for new devices it depends on pyudev
from the python-pyudev package. If this packages is not installed elbe updated
will work as normal but usb monitor functionality is not available. Therfore
the dependency is "only" a suggest.

As the the usb-monitor does not mount the usb device usbmount or similar should
be installed for proper operation.