urllib2.urlopen: use 10 second connect timeout
The default timeout for opening URL's is so long that the elbe client
will timeout before the elbe daemon. Thus the user is left with no
information about why the client timed out. With a smaller timeout
(this patch sets it to 10 seconds), the elbe client is informed that
there is a problem with the URL.

Signed-off-by: John Ogness <john.ogness@linutronix.de>