pbuild: remove existing package from repo before include
The elbe pbuild feature automatically includes the new package to
the elbe local repository. Since a new package with the same
name+suite of an existing package cannot be included in the repo,
commit 78969a6790813c2c57d8c5f96896665a1df27be6 tries to address
this by changing the version of the new package to be a timestamp.
But actually this does not solve the issue since the package name
within the suite has not changed.

With this patch, the automatic version changing is reverted and
any existing packages with the same name+suite are removed from
the local repository before including the new package.

Signed-off-by: John Ogness <john.ogness@linutronix.de>