logging: non-verbose tar (except for archives)
The verbose option was used on all tar commands. This could
create immense output over the soap interface and is not
necessary. The only place where the verbose option is left
enabled is when unpacking project archives. Since this is
globbering any Debian-installed files, there is real value
is including this file list in the log files

Signed-off-by: John Ogness <john.ogness@linutronix.de>