pbuilder: access local repo via http
Some pbuilder packages may depend on previous pbuilder packages
that are only available in the local repository. This can be
accessed via a local http server.

This patch adds a config-package-dev package that will need to
be hosted on the elbe apt source server. This package will
install lighttpd and modify its configuration to run on the
loopback interface port 8080 and only allow access to
/var/cache/elbe/ files.

Signed-off-by: John Ogness <john.ogness@linutronix.de>