use haveged on host for generating enthropy
we route /dev/random into the initvm (by virtio-rng) if
the host doesn't have a hrng, creating gnupg keys inside
the initvm consumes a lot of time. This leads to timeouts
of the 'elbe control' calls that trigger the generation.

to ensure that there's enough enthropy for generating keys
we recommend to install haveged on the host.

this also reverts the installation of the debathena-config-rng-tools
package into the initvm. This package makes the initvm use /dev/urandom
as hrng device. This is maybe not that brilliant, as we thought ;)

Signed-off-by: Manuel Traut <manut@linutronix.de>