cdroms: remove and retry if add package fails
If the package installed in chroot has the same version number but
differs from the package in srcrepo or binrepo, elbe aborts with
an error when the iso images should be generated. Rather than
abort with an error, remove the package from the repo, then try
to add again. If it fails a second time, abort.

This situation can arise if a new pbuilder package is created but
the version number was not incremented.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>