Revert "disable IPv6"
This reverts commit 208ad90a197c17e9f1411eff57901dad150e718b.

The reverted commit is a workaround for a bug in qemu from unstable and
results in a non working 'elbe initvm create' on jessie PCs.

The descirbed bug was not reported to debian, however it should be fixed
there.

Because jessie doesn't use qemu-system 2.7 so this problem doesn't exist if you
don't use stretch or unstable.