improve file uploads into initvm
the upload of the xml file into the initvm was done within one soap
call. so the size of the XML file was limited by soap.

this change removes the old set_xml interface and introduces a new
upload_file interface. the 'elbe control create_project' tool was
also modified, to first create a new empty project and then upload
the xml file.

the project is not allowed to be busy during file uploads. after an
upload, the project gets the status 'has_changes'.

a small function needed to be added to the projectmanager to allow
creating empty projects.

Signed-off-by: Manuel Traut <manut@linutronix.de>