contrib/linux: add missing arch's, remove u+x requirement
The install.sh script for arm, arm64, and powerpc are also modified
so that kernel building support for them also works.

The installkernel script was renamed to elbe_installkernel to be
sure we avoid any nameclashing. This is an elbe-specific change.

The script is now executed by running the interpreter explicitly.
This removes the requirement for the script to be executable. This
is necessary if (for example) the quilt source format is used to
package the kernel. It is also convenient, since tools like "patch"
do not support setting executable bits.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>