Modify minimal stretch example -> network and keyboard support
Modified the stretch example xml file. Previously, it had no X keyboard and
mouse support and the network was not working out of the box.
I added a few Debian packages which configure the above automatically.
To my mind the loss of "minimality" of the example is compensated by a massive
improvement of usability.

Signed-off-by: Philipp Arras <philipp.arras@linutronix.de>