Add/fix version check
It is important that the Elbe versions in initvm and host machine
conincide. The check whether this is the case hasn't been in the correct
position leading to unexpected error messages.

The version is only checked when `elbe control` or `elbe prjrepo` is
called. This is sensible since `elbe initvm` and `elbe pbuilder`
essentially call `elbe control`.

https://github.com/Linutronix/elbe/issues/99

Signed-off-by: Philipp Arras <philipp.arras@linutronix.de>