commands: check_updates: Flush console output before starting script
The "elbe check_updates" sub-command optionally runs a scritp, if a missing or
updatable package is detected. If STDOUT and/or STDERR is redirected to a file
and the script want to access it, no information is contained in this file.

Flushing stdout and stderr before executing the scripts solves this problem.

Signed-off-by: Holger Dengler <dengler@linutronix.de>