chroot: enable local repository
currently the project local repository is not available in 'elbe chroot'
This change makes it available. But take care, if using 'elbe chroot'
during a running 'pbuilder' task. The repo is MOVED into the chroot
and therefore no longer available for pbuilder.

Another fix should use the local webserver to access the repo from
within the chroot.

Signed-off-by: Manuel Traut <manut@linutronix.de>