xml: Add devtmpfs to list of filesystem types
The pseudo filesystem devtmpfs is added to the list of valid fstypes
in XML schema. It provides a list of device nodes, known by the
kernel. It its automount-option is not enabled in kernel config,
devtmpfs must be manually mounted. The automount-option is also
not working in initramfs/initrd, e.g. rescue-filesystems.

Signed-off-by: Holger Dengler <dengler@linutronix.de>