examples: busybox based rescue initramfs with shared libraries
This example creates a busybox based rescue initramfs with shared libs.
It can be used as base for rescue filesystems, which requires further
packages. Adding packages to <pkg-list> will increase the size of the
initramfs only by the size of the package itself and its pre-requisits.

Build size:
  - cpio: 11M
  - cpio.gz: 4.9M

Note:
If the basic glibc built-in iconv/gconv modules are sufficient, the
shared modules in /usr/lib/x86_64-linux-gnu/gconv can be removed. This
will save ~6M (cpio) and ~2.5M (cpio.gz) respectively.

Signed-off-by: Holger Dengler <dengler@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>