validation: use parent directories
To validate repositories, elbe was checking for legacy
per-component-and-architecture Release files. According to
the Debian Repository Format documentation:

    "Clients must not use them."

And indeed, these legacy files are not found on the security
servers.

Instead of the legacy Release files, just check if the parent
directory is there.

NOTE: httpredir.debian.org does not support this. Do not use
      this redirect service with elbe. Validating against a
      redirect service is ridiculous anyway.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>