preseed.cfg: from iso, always install packages from main suite
When installing a root filesytem via network, it is possible that
packages come from other suites (i.e. jessie-backports). However,
when the bin-cdrom.iso is created, all installed initvm packages
are put into the same "main" suite (i.e. jessie).

Originally, when installing via network, it was correct that
debootstrap/jessie-backports was specified in the pkgsel preseed.
But when the same root filesystem is installed via bin-cdrom.iso,
there is no jessie-backports and instead the desired debootstrap
package is in jessie.

When installing from bin-cdrom.iso, strip any non-main suites.
The desired packages will be in the main suite.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>