validation: parse InRelease or Release
With commit 3f77bcb2 "validation: use parent directories" elbe
checks for the parent directory of binary and source information.
However this often fails with redirect services. And although it
it is questionable if validating against a redirect service is a
good idea, the reality of the situation is that most Debian mirrors
are behind redirect services.

Download and parse the InRelease (or if not available, Release)
file for the suite and verify that the appropriate section,
architecture, and source entries are listed.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
(cherry picked from commit 22b491b4905a6bd50ffeeede50c60efc961d36d4)