unset proxy config if not specified in XML
os.environ is persistent for multiple requests to elbe-daemon.
So the proxy configuration of a recent build is still active
for a new build that didn't specify any proxy. This change
sets the proxy env vars to an empty string if no proxy was
specified

This is better than nothing.
However, it should be identified, where these variables are used,
and make sure, that these cases are either isolated in separate
processes (RpcAptCache is a separate process) or called via wget
system or something, to seed the environment when starting the process.

Because this is obviously a race condition, once 2 builds are started
add once on a single initvm.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torbenh@linutronix.de>