virtapt: restrict available archs to system arch
Until now Architectures was not specified. This meant that the
Architectures of the host were used. If the host has added
additional architectures using `dpkg --add-architecture` then
elbe would require these architectures to be available on
mirrors. If the extra unneeded architectures were not on the
mirror, the cache update would fail, causing considerable
confusion for the users.

elbe should not pull such information from the host and certainly
not expect the XML mirrors to provide foreign architectures of
the host.

Explicitly set Architectures to the system architecture. Users
can still add additional architectures via
`dpkg --add-architecture` in finetuning commands.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>