preprocess: allow building variants of a xml file
Other build-systems allow building different flavours or variants of a
image based on some config values. This eases the maintainance of very
similar images.

The elbe preprocess subcommand allows modifying XML files before they are
used with other elbe subcommands. The XML file used by the preprocess
subcommand doesn't need to validate with 'dbsfed.xsd', but the output needs
to.

This adds a new parameter '--variant' to the 'preprocess' subcommand. A
XML tag inside the XML file given to 'preprocess' can contain a 'variant'
attribute. If the value of the variant attribute inside XML matches with
a variant given as parameter to the preprocess command, the XML tag
stays inside the XML file. If the XML tag has a variant attribute but
doesn't match with the variant given as parameter to 'preprocess' the
XML tag will be dropped. If no 'variant' parameter is given to the
'preprocess' subcommand all tags with a 'variant' parameter are dropped.

This allows XML snippets like this:

  <url variant='security'>
          <binary>http://security.debian.org/ stretch/updates main</binary>
          <source>http://security.debian.org/ stretch/updates main</source>
  </url>

  <pkg-list>
          <pkg variant='security'>openssh-server</pkg>
          <pkg variant='audio,video'>totem</pkg>
  </pkg-list>

It is also possible to use multiple sections with variant attributes like this:

<pkg-list variant='audio>
  <pkg>alsa</pkg>
  <pkg>pavucontrol</pkg>
</pkg-list>
<pkg-list variant='video>
  <pkg>totem</pkg>
  <pkg>ffmpeg</pkg>
</pkg-list>

This needs defining mergeable sections by xpath.
A list of mergeable sections is hardcoded in elbepack/xmlpreprocess.py
If one of these sections occure multiple times the contents are merged into a
single section.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>