preseed.cfg: add support for pin and version attributes
To specify packages with an alternate suite, the "pin" attribute
should be used. It may also be desirable for the "version"
attribute to be used. Currently neither of these attributes are
supported in the initvm xml packages.

Add support for both of these attributes for packages specified
in the initvm xml.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by Manuel Traut <manut@linutronix.de>