preferences: use n= for pin attribute
According to the man page for apt_preferences, a= specifies the
suite and n= specifies the codename. elbe documentation discourages
the use of the archive/suite name (i.e. stable, oldstable) because
these change with each release. So since only codenames should be
used (i.e. stretch, jessie), the correct preference setting for the
"pin" attribute would be n=.

Change to use n= for the preference corresponding to the "pin"
attribute.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>