add a build_sdk function
implements build_sdk as a subcommand and as a SOAP function.

It can be used to generate a 'yocto-style' SDK.
Currently only x86_64 is supported as a host sytem and armhf as target.

To generate a SDK use the following commands:

$ elbe initvm submit --keep-files --writeproject uuid.prj examples/armhf-ti-beaglebone-black.xml
$ elbe control build_sdk `cat uuid.prj`
$ elbe control wait_busy `cat uuid.prj`
$ mkdir -p sdk
$ elbe control --output sdk get_files `cat uuid.prj`

To use the SDK, follow the instructions from yocto [0] (Chapter 2)

[0] https://www.yoctoproject.org/docs/2.1/sdk-manual/sdk-manual.html#sdk-using-the-standard-sdk

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>