pkgutils: rework copy_kinitrd function
copy_kinitrd implemented download and extraction of debian packages.
To improve readability and reusability of the code, split out the
download and extraction of a debian package into seperate functions.

Remove code-duplication by introducing a target_dir and dest variable.

Also add optional support for asciidoclog to log into a file instead of
stdout.

Add the posibility to download and extract a specified pacakage
including all its runtime dependencies. (This only works if python-apt
is available)

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>