toolchain-extract.sh: convert to work with mako
Replace ${ by ${"${"} because ${ is a marker of the mako interpreter and $
followed by { is also often used in this shell script.

Also use patchelf that's available at Debian instead of the relocate.py
script from yocto to patch the SDK based location of the libs and
ld-loader into the rpath of host binaries. Because relocate.py requires
the host binaries to be build with a modified gcc/binutils to have some
extra space for longer paths in the ELF header.

utf-8 needs special treatment in mako
(see http://docs.makotemplates.org/en/latest/usage.html#using-unicode-and-encoding)
The utf-8 in an authors name causes an exception:

mako.exceptions.CompileException:
Unicode decode operation of encoding 'ascii' failed in file
'lbepack/makofiles/toolchain-shar-extract.sh.mako' at line: 0 char: 0

Instead of finding the real cause, this removes the utf-8 character from the
file.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>