add SPDX tags to contrib/dockerfile/*
Elbe wants to use SPDX tags in it's code, this part was identified
not to have any copyright statements so far. So this applies the
default: "GPL-3.0+"

But i'm quite unsure about 'contrib/dockerfile/adds/supervisord.conf'
because the original repo of the file is no longer reachable.
Silvio, do you still have access to the copyright informations of this
file?

Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Kate Stewart <kstewart@linuxfoundation.org>