mv get_virtaptcache from rpcaptcache to virtapt
get_virtaptcache is needed by pkgutils. pkgutils.py is in the elbe-common
debian package, but rpcaptcache.py is in the elbe-buildenv debian
package. To avoid moving rpcaptcache.py and all its dependencies into
elbe-common, move the get_virtaptcache function to virtapt.py. This
requires another definition of the 'MyManager' class in virtapt.py.

Signed-off-by: Manuel Traut <manut@linutronix.de>