virtapt: pass keylist to VirtApt
In commit 936e9e2558f1 ("run virtapt in a forked process") the
VirtApt instance is created in a new process. However, the keylist
was not passed to the new forked version of VirtApt. Without the
keylist, elbe would be unable to validate certain repos if the
key was not already imported in the host keyring.

Pass the keylist when creating the VirtApt instance.

Fixes: 936e9e2558f1 ("run virtapt in a forked process")
Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>