Add support for UEFI boot with grub
If the package grub-efi-amd64 is installed and <grub-install/> is
used in the xml, try to install grub on the boot partition
to the default fallback location \EFI\boot\bootx64.efi

Use grub_fw_type to differentiate between "bios" and "efi", so we can
pass that to the grub installer function.

This is currently specific to amd64, but it shouldn't be too difficult
to expand it to other target platforms which support uefi.

Signed-off-by: Martin Kaistra <martin.kaistra@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>