automatically upgrade debootstrap packages
deboostrap runs from the primary_mirror. If a package installed by
deboostrap is available in another repo in a newer version it will not
be installed. However this might be expected.

Adding the debootstrap_pkgs to the target_packages results in the
desired behaviour. Also the correct versions of the packages are listed
in the full_pkg_list.

Reported-by: Lukasz Walewski <lwalewski@s-can.at>
Tested-by: Lukasz Walewski <lwalewski@s-can.at>
Signed-off-by: Manuel Traut <manut@linutronix.de>