soap: implement add_user action
Adding users is only possible via elbe db commands.

When we want to automate management of the initvm, and also
allow sharing of the initvm, its necessary to be able to create
a user via add_user.

This patch exposes ElbeDB.add_user via the soap interface.
add_user is only exposed to an Admin User (mainly root:root)

For our current usecase, we only want to make sure, the user exists.
So an Exception because of an already existing User is ignored, in
the client side code.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>