initvmaction: use systemwide libvirt instance
this change allows several users to share the initvm, because using
the same filesystem image with more than one user results in hard
to debug deadlocking problems with domain start and creation.

This move makes it necessary, that the initvm domain name is configured
to a different name, when more than one initvm shall be running on a
single host.

Due to soap tcp port conflicts, we need special configuration of SOAP
Ports already, so this just requires one more thing that needs to be
configured, when more than one initvm shall run on a single host.

With this patch applied, elbe users need be part of the libvirt group.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>