dbsfed.xsd: reinstate rfs:rm type for <rm> in finetuning.
due to a merge error in
e7bd6e7066c5 ("Merge branch 'bene' into devel/elbe-2.0")
the reference to the rfs:rm complextype which adds the
exclude parameter to the rm finetuning command got removed.

Using the exclude feature resulted in the following error:

----------------------------------------------------------------------------------
error Element 'rm', attribute 'exclude': The attribute 'exclude' is not allowed.
----------------------------------------------------------------------------------

This patch also changes string to rfs:string, because this has not been
changed by b216e05590f5 ("XML schema: allow XML base attribute")
which was on the other side of the merge.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>