projectmanager: allow add_deb_package() to add already existing pkg
includedsc() and includedeb() will raise an Exception, when a pkg
is already in the repo. But we want to be able to add already existing
pkg to the repository, when they are rebuilt.

Use force=True and also determine pkg_name from the filename and
pass that also, because force=True needs the pkgname.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>